move specs from my controller to new access tokens controller
