+++ app/views/my/access_tokens/_storage_tokens_section.html.erb
@@ -62,7 +62,7 @@ See COPYRIGHT and LICENSE files for more details.
                 </td>
                 <td class="buttons">
                   <%= link_to "",
-                              storage_token_delete_path(token.id),
+                              my_access_token_revoke_storage_token_path(token),
                               data: { confirm: t("my_account.access_tokens.storages.revoke_token", storage: token.oauth_client.integration.name),
                                       test_selector: "storages-token-row-#{token.id}-revoke" },
                               method: :delete,
